New Undo System

+++ common/lc_stringcache.cpp
@@ -195,6 +195,6 @@ void lcStringCache::DrawStrings(lcContext* Context, const lcMatrix44* Transforms
 
 	Context->BindTexture2D(mTexture);
 	Context->SetColor(0.0f, 0.0f, 0.0f, 1.0f);
-	Context->DrawPrimitives(GL_TRIANGLES, 0, Strings.size() * 6);
+	Context->DrawPrimitives(GL_TRIANGLES, 0, static_cast<int>(Strings.size()) * 6);
 }
